# swagger.yml
openapi: 3.0.3
info:
  title: GreenCheck Diag Service
  version: 1.1.0
  description: >
    API du microservice de diagnostic GreenCheck.  
    Permet d'analyser l'état de santé d'une plante à partir d'une image et d'un texte utilisateur, 
    et de consulter les diagnostics enregistrés par utilisateur.
servers:
  - url: http://localhost:8000
    description: Serveur local

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    DiagnoseRequest:
      type: object
      required:
        - image_id
        - user_text
      properties:
        image_id:
          type: string
          format: uuid
          description: UUID de l'image stockée sur le cloud
        user_text:
          type: string
          description: Texte saisi par l'utilisateur concernant la plante

    DiagnoseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Identifiant unique du diagnostic
        image_id:
          type: string
          format: uuid
        user_text:
          type: string
        score:
          type: number
          format: float
          description: Note de santé de la plante sur 5
        disease:
          type: string
          description: Maladie détectée, vide si aucune
        advice:
          type: string
          description: Conseils de traitement ou prévention
        created_at:
          type: string
          format: date-time
          description: Date de création du diagnostic
        error:
          type: string
          description: Message d'erreur si applicable

security:
  - BearerAuth: []

paths:
  /diag:
    post:
      summary: Diagnostiquer l'état de santé d'une plante
      description: Lance un diagnostic basé sur une image et un texte utilisateur, et enregistre le résultat en base.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnoseRequest'
      responses:
        '200':
          description: Diagnostic réussi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnoseResponse'
        '400':
          description: 'Mauvaise requête (ex: image non plante)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnoseResponse'
        '401':
          description: JWT invalide ou manquant
        '500':
          description: Erreur interne du serveur
  /diags:
    get:
      summary: Lister tous les diagnostics de l'utilisateur connecté
      description: Retourne la liste complète des diagnostics appartenant à l'utilisateur identifié par le JWT.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des diagnostics de l'utilisateur
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiagnoseResponse'
        '401':
          description: JWT invalide ou manquant

  /diag/{diag_id}:
    get:
      summary: Récupérer un diagnostic spécifique
      description: Retourne le diagnostic correspondant à l'ID, s'il appartient à l'utilisateur connecté.
      security:
        - BearerAuth: []
      parameters:
        - name: diag_id
          in: path
          required: true
          description: Identifiant unique du diagnostic
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Diagnostic trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnoseResponse'
        '401':
          description: JWT invalide ou manquant
        '404':
          description: Diagnostic introuvable ou non autorisé